Clean up slide setup and explain what it does
I changed the strategy to saving local files and then referencing them.
While it technically works to inject the HTML section directly in the
md-page, there were issues with the md to HTML converter that the
framework uses. It could not resolve long nested quotes correctly such as
when including data in the Altair spec, while URLs worked fine.