Fix slide build process and rendering (#176)
* Rename Rmd file to qmd

Easier to automatically render later

* Change to proper revealjs metadata for qmd files

* Set the title in the metadata instead of in the heading

Otherwise all the slide navigation happens vertically
instead of horizontally

* Add sliding animation

* Use knitr engine

By default the python engine is used when there are only python files
but rendering with knitr giver more control over
both pandas tables and altair charts.

* Use aside class for transcript

* Delete html file which is now dynamically generated

* Use dynamically generated files

The Html version of these are automatically added to the built site

* Switch to robust solution for showing speaker notes

This will be more reliable than a custom div
that we need to position manually.

* Increase space that slides/video occupy

Since the layout of the slide pages is "full"
they will occupy the margin as well
when there is no ToC.
This is why it is useful to also increase the margin-width.

* Set dimensions to be the same for slides and video

* Add slide numbers and ensure slides take up full width

* Include speaker notes for title slide